refactor: Separate canvas from steps widgets and move to visualization. Move plots to visualization.
