+++ src/visualization/__init__.py
@@ -1,1 +1,3 @@
 from .a_priori_canvas import APrioriGauge, APrioriGraphPlot, APrioriScatterPlot
+from .gmm_canvas import GMMCanvas
+from .k_means_canvas import KMeansCanvas
