+++ src/widgets/steps_widgets/__init__.py
@@ -1,5 +1,5 @@
 from .algorithm_vis import AlgorithmStepsVisualization
 from .a_priori_vis import APrioriStepsVisualization
 from .extra_trees_vis import ExtraTreesStepsVisualization
-from .gmm_vis import GMMStepsVisualization, GMMCanvas
-from .k_means_vis import KMeansStepsVisualization, KMeansCanvas
+from .gmm_vis import GMMStepsVisualization
+from .k_means_vis import KMeansStepsVisualization
